Split test groups into different files
